# Dynamically bypass selector polling if no I/O events are present