--- a/core/jvm/src/main/scala/cats/effect/unsafe/WorkStealingThreadPool.scala
+++ b/core/jvm/src/main/scala/cats/effect/unsafe/WorkStealingThreadPool.scala
@@ -47,6 +47,7 @@ import java.util.concurrent.atomic.{
   AtomicReference,
   AtomicReferenceArray
 }
+import java.util.concurrent.locks.LockSupport
 
 import WorkStealingThreadPool._
 
@@ -93,7 +94,8 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
     new AtomicReferenceArray(threadCount)
   private[unsafe] val localQueues: Array[LocalQueue] = new Array(threadCount)
   private[unsafe] val sleepers: Array[TimerHeap] = new Array(threadCount)
-  private[unsafe] val parkedSignals: Array[AtomicBoolean] = new Array(threadCount)
+  private[unsafe] val parkedSignals: Array[AtomicReference[ParkedSignal]] = new Array(
+    threadCount)
   private[unsafe] val fiberBags: Array[WeakBag[Runnable]] = new Array(threadCount)
   private[unsafe] val pollers: Array[P] =
     new Array[AnyRef](threadCount).asInstanceOf[Array[P]]
@@ -151,7 +153,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
       localQueues(i) = queue
       val sleepersHeap = new TimerHeap()
       sleepers(i) = sleepersHeap
-      val parkedSignal = new AtomicBoolean(false)
+      val parkedSignal = new AtomicReference[ParkedSignal](ParkedSignal.Unparked)
       parkedSignals(i) = parkedSignal
       val index = i
       val fiberBag = new WeakBag[Runnable]()
@@ -241,7 +243,7 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
 
       if (isStackTracing) {
         destWorker.active = fiber
-        parkedSignals(dest).lazySet(false)
+        parkedSignals(dest).lazySet(ParkedSignal.Unparked)
       }
 
       fiber
@@ -306,14 +308,12 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
       val index = (from + i) % threadCount
 
       val signal = parkedSignals(index)
-      if (signal.getAndSet(false)) {
-        // Update the state so that a thread can be unparked.
-        // Here we are updating the 16 most significant bits, which hold the
-        // number of active threads, as well as incrementing the number of
-        // searching worker threads (unparked worker threads are implicitly
-        // allowed to search for work in the local queues of other worker
-        // threads).
-        state.getAndAdd(DeltaSearching)
+      val st = signal.get()
+
+      if ((st eq ParkedSignal.ParkedPolling) || (st eq ParkedSignal.ParkedSimple) || signal
+          .compareAndSet(st, ParkedSignal.Interrupting)) {
+        doneSleeping()
+
         // Fetch the latest references to the worker threads before doing the
         // actual unparking. There is no danger of a race condition where the
         // parked signal has been successfully marked as unparked but the
@@ -322,7 +322,14 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
         // point it is already unparked and entering this code region is thus
         // impossible.
         val worker = workerThreads.get(index)
-        system.interrupt(worker, pollers(index))
+
+        if (st eq ParkedSignal.ParkedPolling) {
+          system.interrupt(worker, pollers(index))
+        } else {
+          LockSupport.unpark(worker)
+        }
+        signal.set(ParkedSignal.Unparked)
+
         return true
       }
 
@@ -446,6 +453,12 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
   }
 
   private[unsafe] def doneSleeping(): Unit = {
+    // Update the state so that a thread can be unparked.
+    // Here we are updating the 16 most significant bits, which hold the
+    // number of active threads, as well as incrementing the number of
+    // searching worker threads (unparked worker threads are implicitly
+    // allowed to search for work in the local queues of other worker
+    // threads).
     state.getAndAdd(DeltaSearching)
     ()
   }
@@ -710,6 +723,8 @@ private[effect] final class WorkStealingThreadPool[P <: AnyRef](
       while (i < threadCount) {
         val workerThread = workerThreads.get(i)
         if (workerThread ne currentThread) {
+          // we don't know which state we're in, so just try both interruptions
+          LockSupport.unpark(workerThread)
           system.interrupt(workerThread, pollers(i))
           workerThread.interrupt()
         }