@djspiewak 2025-04-13
Released as `3.6.1-25-d807be0` if anyone would like to test something