@djspiewak 2025-04-14
`DispatcherBenchmark` is the biggest change, regressing 25%. `parTraverse` regressed slightly but so did `runnableSchedulingScalaGlobal`, which is more of a concurrency baseline, and `parTraverse` regressed *less* so let's call that one a tentative win. `traverse` is equivalent (a non-concurrency baseline). All the main WSTP benchmarks are essentially within the margin of error.

I'd like to see the real world numbers on this one, but the microbenchmarks seem to give their blessing in so far as they go.

## Before

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score     Error    Units
[info] DispatcherBenchmark.scheduling                                N/A     1000  thrpt   20   169.751 ± 100.207  ops/min
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   20   543.307 ±   1.551    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   20    50.845 ±   0.015    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   20     9.510 ±   0.049  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   20    12.188 ±   0.822  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   20  1529.989 ±  38.601  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   20  1746.552 ±   9.345  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   20    13.742 ±   0.398  ops/min
```

## After

Note that the global EC is 4% slower in this run, but `traverse` is essentially within the margin of error. That indicates that we would generally expect anything concurrency-related to have a lower baseline in this run than in the previous one. Good old EC2!

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score    Error    Units
[info] DispatcherBenchmark.scheduling                                N/A     1000  thrpt   20   125.216 ± 15.753  ops/min
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   20   529.290 ± 26.988    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   20    50.864 ±  0.008    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   20     9.499 ±  0.036  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   20    12.800 ±  0.604  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   20  1539.262 ± 13.460  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   20  1683.373 ± 25.941  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   20    13.744 ±  0.501  ops/min
```