@djspiewak 2025-04-14
So my interpretation of the results from #4328 is that this is indeed an improvement, and we likely have some more work to do on `Selector` constant factor overhead.