@durban 2025-04-14
I'm probably missing something, but do we _really_ need the `ParkedSignal.Interrupting` state? What does the spinlock _actually_ protects? (Spurious wakeups already must be handled.)

(Btw, I've restarted the "windows-latest, 2.13.15, temurin@17, ciJVM" job, because it timed out. But I'm wondering if it was a real failure.)