@durban 2025-04-14
Okay, so the way to wake someone up is (without this PR) essentially:
```scala
if (signal.getAndSet(false)) system.interrupt(...)
```

With the 2 different parking states it could be:
```scala
val old = signal.getAndSet(Unparked)
// HERE
if (old eq ParkedPolling) system.interrupt(...)
else if (old eq ParkedSimple) LockSupport.unpark(...)
```

And yes, it could happen that at the point marked with `HERE` above the target thread wakes up, then goes back to sleep with a different way. But is that a problem? We wanted to wake it up, and... it did! (And sure, then we might make a wrong unpark call, but that should be harmless(?).)

(Again, I might be missing something, I just started thinking about this...)