@iRevive 2025-04-17
I tried `3.6.1-25-d807be0`, here are the results:

### CPU Usage

<img width="1822" alt="image" src="https://github.com/user-attachments/assets/5fc1f4d8-591c-44da-a7f5-82f50c87c872" />

Although the graph shows the mean value as 20%, the startup inflates the number. The more realistic number is ~13% when the system is idling.

### Events per worker

<img width="1822" alt="image" src="https://github.com/user-attachments/assets/f80dc93a-4737-4f83-b32d-e4531f58709b" />

The sum of all events (blocked, parked, polled) across all workers. 
For instance, CE 3.6.1 SelectorSystem polls the thread **7.690.942** times per 5 minutes, CE `3.6.1-25-d807be0` - **11.766**.